Added a section about deploying the tests on EKS
This section should prompt any developers working on ginkgo tests that the image tag needs to be updated in govuk-helm-charts repo as it is not automatically deployed like other govuk apps.

There is an additional note about how to manually trigger the ECR pulll through cache as it is not currently working for this repo. An issue has been written up describing the problem. This section should be removed once the issue has been fixed.